Secure views on book admin

+++ books/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -61,6 +62,7 @@ def all_books(request):
 
     return render(request, 'books/books.html', context)
 
+
 def book_detail(request, book_id):
     """ A view to show individual book details """
     book = get_object_or_404(Book, pk=book_id)
@@ -72,8 +74,13 @@ def book_detail(request, book_id):
     return render(request, 'books/book_detail.html', context)
 
 
+@login_required
 def add_book(request):
     """ Add a book to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, unauthorised access. Site owner access only')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = BookForm(request.POST, request.FILES)
         if form.is_valid():
@@ -93,8 +100,13 @@ def add_book(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_book(request, book_id):
     """ Edit a book in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, unauthorised access. Site owner access only')
+        return redirect(reverse('home'))
+
     book = get_object_or_404(Book, pk=book_id)
     if request.method == 'POST':
         form = BookForm(request.POST, request.FILES, instance=book)
@@ -117,8 +129,13 @@ def edit_book(request, book_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_book(request, book_id):
     """ Delete a book from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, unauthorised access. Site owner access only')
+        return redirect(reverse('home'))
+        
     book = get_object_or_404(Book, pk=book_id)
     book.delete()
     messages.success(request, 'Book deleted!')
